Remove ChartData as component
